# contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: WSC Access Workbench API
  version: 2.2.0
  description: |
    wsc-contracts@2.2.0 — 接入端工作台 V1.4（座序图 L2 Top5、持久化用户、我的数据产品；SNAP-WSC-005）。
    统一响应包装：{ code, message, data, correlationId }；成功时 code="0"。
    RBAC：产品增/改/导入仅 PROVIDER；ADMIN 保留分类维护与用户管理；目录维护 ADMIN+PROVIDER（PROVIDER 仅 create_by=本人）。
    GET /catalog/products?mine=true 仅返回 create_by=当前用户的产品。
    GET /catalog/l2-distribution 返回二级目录产品计数（座序图）。
    ADMIN /admin/users CRUD 持久化 sys_user。
    产品编码格式（DEC-WSC-001）：`{DOMAIN}-{FEATURE}-{NNNN}`。
    认证：服务端会话（Cookie）；写 API 以服务端会话角色为准。
    （HOTFIX TASK-WSC-607：目录维护对 PROVIDER 开放本人产品；契约版本号未 bump，仍为 2.2.0。）
servers:
  - url: /api/v1
    description: V1 API base
...
security:
  - SessionCookie: []
paths:
  /auth/session:
    get:
      tags: [auth]
      summary: 获取当前会话
      operationId: getSession
      responses:
        "200":
          description: 当前会话
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseSession"
              examples:
                ok:
                  value:
                    code: "0"
                    message: OK
                    data:
                      userId: "u-1"
                      displayName: "演示用户"
                      role: ADMIN
                      enterpriseName: "演示企业"
                    correlationId: "corr-auth-1"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [auth]
      summary: 登录建立会话
      operationId: createSession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: 登录成功（Set-Cookie 会话）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseSession"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [auth]
      summary: 登出（立即失效会话）
      operationId: deleteSession
      responses:
        "200":
          description: 已登出
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseEmpty"
  /auth/session/role:
    post:
      tags: [auth]
      summary: 会话角色切换（始终禁用）
      description: |
        V1.4 起角色仅由 sys_user 账号绑定；任意会话角色切换一律禁用。
        三角色（ADMIN/PROVIDER/USER）调用均返回 HTTP 410 + ERR_ROLE_SWITCH_DISABLED。
        本操作无成功路径。
      operationId: switchSessionRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchRoleRequest"
      responses:
        "410":
          description: 角色切换已永久禁用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              examples:
                disabled:
                  summary: 任意角色均禁用
                  value:
                    code: ERR_ROLE_SWITCH_DISABLED
                    message: 角色由账号绑定，禁止会话切换
                    data: null
                    correlationId: "corr-role-switch-410"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /overview/trend:
    get:
      tags: [overview]
      summary: 近 30 天上链趋势（REQ-OVW-002）
      operationId: getOverviewTrend
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
            minimum: 1
            maximum: 90
      responses:
        "200":
          description: 趋势序列
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseOverviewTrend"
  /overview/top:
    get:
      tags: [overview]
      summary: 产品上链次数 TOP（REQ-OVW-003）
      operationId: getOverviewTop
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 10
      responses:
        "200":
          description: TOP 列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseOverviewTop"
  /overview/stream:
    get:
      tags: [overview]
      summary: 最新上链动态流（REQ-OVW-004）
      operationId: getOverviewStream
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: 动态流（V1 来自投影/fixture，不依赖未实现业务 API）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseOverviewStream"
...
  /admin/users:
    get:
      tags: [admin]
      summary: 用户列表（仅 ADMIN；REQ-USER-001）
      description: |
        响应为命名 schema AdminUser[]；禁止回显 password / passwordHash。
        非 ADMIN → ERR_FORBIDDEN（403）。
      operationId: listUsers
      responses:
        "200":
          description: 用户列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAdminUserPage"
              examples:
                ok:
                  value:
                    code: "0"
                    message: OK
                    data:
                      items:
                        - userId: "u-1"
                          username: "admin"
                          displayName: "管理员"
                          role: ADMIN
                          enterpriseName: "演示企业"
                          deleted: false
                    correlationId: "corr-users-list"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [admin]
      summary: 创建用户（仅 ADMIN；REQ-USER-001）
      description: |
        必填 username + password + role；可选 displayName / enterpriseName。
        username 冲突 → ERR_USER_USERNAME_CONFLICT（409）。
        响应 AdminUser 禁止含 password / passwordHash。
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserCreate"
      responses:
        "200":
          description: 已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAdminUser"
              examples:
                ok:
                  value:
                    code: "0"
                    message: OK
                    data:
                      userId: "u-2"
                      username: "provider1"
                      displayName: "提供方一"
                      role: PROVIDER
                      enterpriseName: "演示企业"
                      deleted: false
                    correlationId: "corr-users-create"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: username 冲突
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              examples:
                conflict:
                  value:
                    code: ERR_USER_USERNAME_CONFLICT
                    message: 用户名已存在
                    data: null
                    correlationId: "corr-users-dup"
  /admin/users/{userId}:
    put:
      tags: [admin]
      summary: 更新用户角色/资料或软删（仅 ADMIN；REQ-USER-001）
      description: |
        可改角色、资料、密码；软删权威形状为 PUT deleted: true（本版不采用 DELETE）。
        用户不存在 → ERR_USER_NOT_FOUND（404）。
        响应 AdminUser 禁止含 password / passwordHash。
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
            examples:
              softDelete:
                summary: 软删
                value:
                  deleted: true
              changeRole:
                summary: 改角色
                value:
                  role: USER
      responses:
        "200":
          description: 已更新
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAdminUser"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              examples:
                notFound:
                  value:
                    code: ERR_USER_NOT_FOUND
                    message: 用户不存在
                    data: null
                    correlationId: "corr-users-404"
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: WSC_SESSION
      description: |
        服务端会话 Cookie。登录后写入；登出立即失效。
        角色绑定在会话中；写 API 以服务端会话角色为准。
        角色/权限变更后须吊销或强制刷新旧会话。
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-or-JWT
      description: |
        可选等价方案：Bearer 令牌必须映射可吊销的服务端会话表。
        非「纯无状态 JWT」；吊销后写 API 必须拒绝。
...
  responses:
    Unauthorized:
      description: 未认证
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponseError"
          examples:
            unauth:
              value:
                code: "401"
                message: 未登录或会话已失效
                data: null
                correlationId: "corr-401"
    Forbidden:
      description: 无权限
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponseError"
          examples:
            forbidden:
              value:
                code: ERR_FORBIDDEN
                message: 当前角色无权执行该写操作
                data: null
                correlationId: "corr-403"
...
  schemas:
    ApiEnvelope:
      type: object
      required: [code, message, data, correlationId]
      properties:
        code:
          type: string
          description: 成功为 "0"；业务错误为 ERR_* 或 HTTP 语义码字符串
        message:
          type: string
        data:
          nullable: true
        correlationId:
          type: string
    ApiResponseEmpty:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              nullable: true
    ApiResponseError:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            code:
              type: string
              enum:
                - ERR_PRODUCT_CODE_FORMAT
                - ERR_PRODUCT_CODE_CONFLICT
                - ERR_CATEGORY_HAS_PRODUCTS
                - ERR_CATEGORY_LEAF_REQUIRED
                - ERR_MAINTENANCE_FORBIDDEN
                - ERR_IMPORT_FILE_TOO_LARGE
                - ERR_IMPORT_FORMAT_INVALID
                - ERR_IMPORT_TEMPLATE_UNSUPPORTED
                - ERR_IMPORT_REPORT_NOT_FOUND
                - ERR_FORBIDDEN
                - ERR_ROLE_SWITCH_DISABLED
                - ERR_USER_USERNAME_CONFLICT
                - ERR_USER_NOT_FOUND
                - "401"
                - "404"
            data:
              nullable: true
              description: ERR_CATEGORY_HAS_PRODUCTS 时含 reason、productCount
    Role:
      type: string
      enum: [ADMIN, PROVIDER, USER]
...
    LoginRequest:
      type: object
      required: [username, password]
      description: |
        登录仅接受 username + password。角色不得由客户端选择；
        会话角色仅来自 sys_user 账号绑定（V1.4 废止 LoginRequest.role）。
      properties:
        username: { type: string }
        password: { type: string, format: password }
    SwitchRoleRequest:
      type: object
      required: [role]
      description: 会话切角色请求体（端点始终 410；无成功路径）
      properties:
        role:
          $ref: "#/components/schemas/Role"
    Session:
      type: object
      required: [userId, displayName, role]
      properties:
        userId: { type: string }
        displayName: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        enterpriseName: { type: string }
    AdminUser:
      type: object
      required: [userId, username, displayName, role, enterpriseName, deleted]
      description: |
        用户管理响应模型。list/get/create/update 响应禁止出现 password / passwordHash。
      properties:
        userId: { type: string }
        username: { type: string }
        displayName: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        enterpriseName: { type: string }
        deleted: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AdminUserCreate:
      type: object
      required: [username, password, role]
      properties:
        username: { type: string }
        password: { type: string, format: password }
        role: { $ref: "#/components/schemas/Role" }
        displayName: { type: string }
        enterpriseName: { type: string }
    AdminUserUpdate:
      type: object
      description: |
        更新用户；软删权威形状为 deleted: true（本版不采用 DELETE）。
        password 仅作写输入，响应永不回显。
      properties:
        displayName: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        enterpriseName: { type: string }
        password: { type: string, format: password }
        deleted: { type: boolean }
    AdminUserPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
    ApiResponseAdminUser:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/AdminUser"
    ApiResponseAdminUserPage:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/AdminUserPage"
...
    ApiResponseSession:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Session"
...
    TrendPoint:
      type: object
      required: [date, count]
      properties:
        date: { type: string, format: date }
        count: { type: integer }
    OverviewTrend:
      type: object
      required: [points]
      properties:
        points:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
    ApiResponseOverviewTrend:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/OverviewTrend"
    TopItem:
      type: object
      required: [productName, count]
      properties:
        productId: { type: string }
        productName: { type: string }
        count: { type: integer }
    OverviewTop:
      type: object
      required: [items]
      properties:
        items:
          type: array
          maxItems: 10
          items:
            $ref: "#/components/schemas/TopItem"
    ApiResponseOverviewTop:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/OverviewTop"
    StreamEventType:
      type: string
      enum: [CATALOG_REGISTER, DATA_REGISTER, TRADE_ORDER]
    StreamEvent:
      type: object
      required: [type, subject, actionSummary, relativeTime, occurredAt, chainRecordId]
      properties:
        type: { $ref: "#/components/schemas/StreamEventType" }
        subject: { type: string }
        actionSummary: { type: string }
        relativeTime: { type: string }
        occurredAt: { type: string, format: date-time }
        chainRecordId: { type: string }
    OverviewStream:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/StreamEvent"
    ApiResponseOverviewStream:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/OverviewStream"
...